accessible/tests/mochitest/actions/test_tree.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL tree actions tests">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../treeview.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../states.js" />
  <script type="application/javascript"
          src="../events.js" />
  <script type="application/javascript"
          src="../actions.js" />

  <script type="application/javascript">
  <![CDATA[
    /* global testAction */

    async function doTest()
    {
      var treeNode = getNode("tree");

      var treeBodyNode = treeNode.treeBody;

      var tree = getAccessible(treeNode);
      var expandedTreeItem = tree.getChildAt(2);
      var collapsedTreeItem = tree.getChildAt(5);

      await testAction({
        id: expandedTreeItem,
        actionName: "activate",
        actionIndex: 0,
        events: CLICK_EVENTS,
        targetId: treeBodyNode,
        eventSeq: [
          [EVENT_FOCUS, expandedTreeItem,
           () => testStates(expandedTreeItem, STATE_FOCUSED | STATE_SELECTED | STATE_EXPANDED)]
        ]
      });

      await testAction({
        id: collapsedTreeItem,
        actionName: "expand",
        actionIndex: 1,
        events: CLICK_EVENTS,
        targetId: treeBodyNode
      });
      testStates(collapsedTreeItem, STATE_EXPANDED);

      await testAction({
        id: collapsedTreeItem,
        actionName: "collapse",
        actionIndex: 1,
        events: CLICK_EVENTS,
        targetId: treeBodyNode
      });
      testStates(collapsedTreeItem, STATE_COLLAPSED);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yXULTreeLoadEvent(doTest, "tree", new nsTreeTreeView());
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank" rel="opener"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=503727"
         title="Reorganize implementation of XUL tree accessibility">
        Mozilla Bug 503727
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <tree id="tree" flex="1" minheight="100px">
        <treecols>
          <treecol id="col" flex="1" primary="true" label="column"/>
        </treecols>
        <treechildren/>
      </tree>

      <vbox id="debug"/>
    </vbox>
  </hbox>

</window>
